[WIP] More prototypejs -> jQuery transition.
So it seems the concept of fully isolated objects is
actually more like a nightmare.

Though still inline with previous plan. I forced to
expose our post object and hence I attach it to Moebooru.
I hope it reasonable enough, since Moebooru actually being used as
a facade now.

Here attached stub for vote.js which will replace the previous vote-widget.
Most of the code still a mess, and likely to change.

Some pattern will be (hopefully) kept as is:
    - Moebooru/Moe is the only global object we will use throughout the apps
      no other globals may be used (except ext. library).
    - all custom events will use this naming convention:
      <objname>:<methodname>[:<status>]
      e.g. post:add
           vote:set:error
    - for API call (Moebooru.request) any callback should listen to this event
      <API url>:<status>
      e.g. /post/vote.json:ready